Implement Race Function
Overview
----
* Added Promise.race() to the library to allow using it.
* Added test scripts for Promise.race()
* Updated README.md to show .race use

Use Cases
----
Using within Qubit experiences, allows developers to have two or more ways of having the experience run.

```js
Promise.race([waitForBasketSummary(), waitForBasketAdd()])
  .then(pollForMiniCart)
  .then(fire)
```

In the above example, waitForBasketSummary sets up a `options.uv.on` while waitForBasketAdd sets up an `ajaxComplete`. This allows it to run if either the uv.on fires or if the
ajaxComplete fires. Whichever happens first, but either running is fine to happen.